rename sum_tree.py to binary_tree.py
